# Empty repositories should not have dirty indications